requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
babel==2.17.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
datastar-py==0.5.0
Django==5.2.3
django-debug-toolbar==5.2.0
django-extra-views==0.15.0
django-haystack==3.3.0
django-oscar==4.0
django-phonenumber-field==8.1.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-tables2==2.7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
django_good_otp==2.0.4
factory_boy==3.3.3
Faker==37.3.0
frozenlist==1.7.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
multidict==6.5.0
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
purl==1.6
pycountry==24.6.1
PyJWT==2.10.1
pyotp==2.9.0
pysolr==3.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qrcode==8.2
redis==6.2.0
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.3
twilio==9.6.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Whoosh==2.7.4
yarl==1.20.1
